requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Data Visualization
plotly>=5.15.0

# API and HTTP
requests>=2.31.0

# Configuration Management
python-dotenv>=1.0.0

# PDF Generation
reportlab>=4.0.0

# Optional: Time Series Forecasting
# Uncomment if using forecast functionality
# prophet>=1.1.4
# statsmodels>=0.14.0

# Optional: Scientific Computing
# Uncomment if using imsciences module
# imsciences

# Development Dependencies (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
# pytest>=7.4.0